changes to roll.js
